moved component to respective folders
